Move oidc tests into oidc_feature.go files
