+++ test/rekt/broker_test.go
@@ -20,6 +20,7 @@ limitations under the License.
 package rekt
 
 import (
+	"knative.dev/reconciler-test/pkg/feature"
 	"testing"
 	"time"
 
@@ -33,7 +34,7 @@ import (
 
 	"knative.dev/eventing/test/rekt/features/broker"
 	"knative.dev/eventing/test/rekt/features/oidc"
-	b "knative.dev/eventing/test/rekt/resources/broker"
+	brokerresources "knative.dev/eventing/test/rekt/resources/broker"
 )
 
 func TestBrokerWithManyTriggers(t *testing.T) {
@@ -81,7 +82,7 @@ func TestBrokerAsMiddleware(t *testing.T) {
 	)
 
 	// Install and wait for a Ready Broker.
-	env.Prerequisite(ctx, t, broker.GoesReady("default", b.WithEnvConfig()...))
+	env.Prerequisite(ctx, t, broker.GoesReady("default", brokerresources.WithEnvConfig()...))
 
 	// Test that a Broker can act as middleware.
 	env.Test(ctx, t, broker.SourceToSink("default"))
@@ -120,7 +121,7 @@ func TestBrokerWithFlakyDLQ(t *testing.T) {
 	)
 
 	// Install and wait for a Ready Broker.
-	env.Prerequisite(ctx, t, broker.GoesReady("default", b.WithEnvConfig()...))
+	env.Prerequisite(ctx, t, broker.GoesReady("default", brokerresources.WithEnvConfig()...))
 
 	// Test that a Broker can act as middleware.
 	env.Test(ctx, t, broker.SourceToSinkWithFlakyDLQ("default"))
@@ -138,9 +139,9 @@ func TestBrokerConformance(t *testing.T) {
 	)
 
 	// Install and wait for a Ready Broker.
-	env.Prerequisite(ctx, t, broker.GoesReady("default", b.WithEnvConfig()...))
+	env.Prerequisite(ctx, t, broker.GoesReady("default", brokerresources.WithEnvConfig()...))
 	env.TestSet(ctx, t, broker.DataPlaneConformance("default"))
-	env.TestSet(ctx, t, broker.ControlPlaneConformance("default", b.WithEnvConfig()...))
+	env.TestSet(ctx, t, broker.ControlPlaneConformance("default", brokerresources.WithEnvConfig()...))
 }
 
 func TestBrokerDefaultDelivery(t *testing.T) {
@@ -263,9 +264,9 @@ func TestBrokerSupportsOIDC(t *testing.T) {
 	)
 
 	name := feature.MakeRandomK8sName("broker")
-	env.Prerequisite(ctx, t, brokerfeatures.GoesReady(name, broker.WithEnvConfig()...))
+	env.Prerequisite(ctx, t, broker.GoesReady(name, brokerresources.WithEnvConfig()...))
 
-	env.TestSet(ctx, t, oidc.AddressableOIDCConformance(broker.GVR(), "Broker", name, env.Namespace()))
+	env.TestSet(ctx, t, oidc.AddressableOIDCConformance(brokerresources.GVR(), "Broker", name, env.Namespace()))
 }
 
 func TestBrokerSendsEventsWithOIDCSupport(t *testing.T) {
