+++ test/rekt/features/containersource/features.go
@@ -56,38 +56,6 @@ func SendsEventsWithSinkRef() *feature.Feature {
 	return f
 }
 
-func SendsEventsWithSinkRefOIDC() *feature.Feature {
-	src := feature.MakeRandomK8sName("containersource")
-	sink := feature.MakeRandomK8sName("sink")
-	sinkAudience := "audience"
-	f := feature.NewFeature()
-
-	f.Prerequisite("OIDC authentication is enabled", featureflags.AuthenticationOIDCEnabled())
-	f.Prerequisite("transport encryption is strict", featureflags.TransportEncryptionStrict())
-	f.Prerequisite("should not run when Istio is enabled", featureflags.IstioDisabled())
-
-	f.Setup("install sink", eventshub.Install(sink,
-		eventshub.OIDCReceiverAudience(sinkAudience),
-		eventshub.StartReceiverTLS))
-
-	f.Requirement("install ContainerSource", func(ctx context.Context, t feature.T) {
-		d := service.AsDestinationRef(sink)
-		d.CACerts = eventshub.GetCaCerts(ctx)
-		d.Audience = &sinkAudience
-
-		containersource.Install(src, containersource.WithSink(d))(ctx, t)
-	})
-
-	f.Requirement("containersource goes ready", containersource.IsReady(src))
-
-	f.Stable("containersource as event source").
-		Must("delivers events",
-			assert.OnStore(sink).MatchEvent(test.HasType("dev.knative.eventing.samples.heartbeat")).AtLeast(1)).
-		Must("Set sinkURI to HTTPS endpoint", source.ExpectHTTPSSink(containersource.Gvr(), src)).
-		Must("Set sinkCACerts to non empty CA certs", source.ExpectCACerts(containersource.Gvr(), src))
-	return f
-}
-
 func SendsEventsWithSinkURI() *feature.Feature {
 	source := feature.MakeRandomK8sName("containersource")
 	sink := feature.MakeRandomK8sName("sink")
