+++ test/rekt/features/pingsource/features.go
@@ -265,34 +265,3 @@ func SendsEventsWithBrokerAsSinkTLS() *feature.Feature {
 	return f
 
 }
-
-func PingSourceSendEventOIDC() *feature.Feature {
-	source := feature.MakeRandomK8sName("pingsource")
-	sink := feature.MakeRandomK8sName("sink")
-	sinkAudience := "audience"
-	f := feature.NewFeature()
-
-	f.Prerequisite("OIDC authentication is enabled", featureflags.AuthenticationOIDCEnabled())
-	f.Prerequisite("transport encryption is strict", featureflags.TransportEncryptionStrict())
-	f.Prerequisite("should not run when Istio is enabled", featureflags.IstioDisabled())
-
-	f.Setup("install sink", eventshub.Install(sink,
-		eventshub.OIDCReceiverAudience(sinkAudience),
-		eventshub.StartReceiverTLS))
-
-	f.Requirement("Install pingsource", func(ctx context.Context, t feature.T) {
-		d := service.AsDestinationRef(sink)
-		d.CACerts = eventshub.GetCaCerts(ctx)
-		d.Audience = &sinkAudience
-
-		pingsource.Install(source, pingsource.WithSink(d))(ctx, t)
-	})
-
-	f.Requirement("pingsource goes ready", pingsource.IsReady(source))
-
-	f.Stable("pingsource as event source").
-		Must("delivers events",
-			assert.OnStore(sink).MatchEvent(test.HasType("dev.knative.sources.ping")).AtLeast(1))
-
-	return f
-}
